# mkdocs-site/docs/api/specs/openapi.yaml
openapi: 3.1.0
info:
  title: Baselith-Core
  version: 0.1.0
paths:
...
  /feedback:
    post:
      tags:
      - feedback
      summary: Feedback
      description: 'Records a feedback (positive|negative) for a generated response.

        - Data is saved to the PostgreSQL database (configurable via the dedicated
        environment variables).'
      operationId: feedback_feedback_post
      requestBody:
        content:
          application/json:
            schema:
              additionalProperties: true
              type: object
              title: Payload
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
                title: Response Feedback Feedback Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /feedbacks:
    get:
      tags:
      - feedback
      summary: List Feedbacks
      description: 'Returns all saved feedback entries.

        - If `feedback` is specified, filters by type (''positive'' or ''negative'').'
      operationId: list_feedbacks_feedbacks_get
      parameters:
      - name: feedback
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - positive
            - negative
            type: string
          - type: 'null'
          description: 'Filter results by feedback type: ''positive'' or ''negative'''
          title: Feedback
        description: 'Filter results by feedback type: ''positive'' or ''negative'''
      - name: limit
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 200
            minimum: 1
          - type: 'null'
          description: Limit the number of returned records (max 200)
          title: Limit
        description: Limit the number of returned records (max 200)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Feedbacks Feedbacks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin:
    get:
      tags:
      - admin
      summary: Admin Page
      description: Restituisce la pagina Admin (protetta da Basic Auth).
      operationId: admin_page_admin_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBasic: []
  /admin/data:
    get:
      tags:
      - admin
      summary: Admin Data
      description: 'Returns analytics/feedback data in JSON format (protected by Basic
        Auth).

        - Aggregated totals (positive, negative, percentages)

        - Daily time series

        - Recent feedback with metadata and sources

        - Most cited queries and documents in the selected period'
      operationId: admin_data_admin_data_get
      security:
      - HTTPBasic: []
      parameters:
      - name: days
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
            maximum: 365
            minimum: 1
          - type: 'null'
          description: Time window (in days) to consider for analytics. Use values
            >0.
          default: 30
          title: Days
        description: Time window (in days) to consider for analytics. Use values >0.
      - name: recent_limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: Number of recent feedbacks to return.
          default: 20
          title: Recent Limit
        description: Number of recent feedbacks to return.
      - name: top_limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          description: Maximum number of entries for popular queries/documents.
          default: 10
          title: Top Limit
        description: Maximum number of entries for popular queries/documents.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /admin/tenants:
    get:
      tags:
      - admin
      - tenants
      summary: List Tenants
      description: List all tenants.
      operationId: list_tenants_admin_tenants_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Tenant'
                type: array
                title: Response List Tenants Admin Tenants Get
      security:
      - HTTPBasic: []
    post:
      tags:
      - admin
      - tenants
      summary: Create Tenant
      description: Create a new tenant.
      operationId: create_tenant_admin_tenants_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBasic: []
components:
  schemas:
...
    CreateTenantRequest:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: CreateTenantRequest
      description: Payload for creating a new tenant.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Tenant:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        status:
          type: string
          title: Status
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - id
      - name
      - status
      - created_at
      title: Tenant
      description: "Data model representing a system tenant.\n\nAttributes:\n    id:\
        \ Unique tenant identifier (slug).\n    name: Human-readable display name.\n\
        \    status: Operational status (active, suspended, etc.).\n    created_at:\
        \ ISO timestamp of creation."
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBasic:
      type: http
      scheme: basic